Fixed my shit computer

+++ src/com/aidenlauris/gameobjects/Enemy.java
@@ -22,7 +22,7 @@ public class Enemy extends Entity {
 	private boolean isHovering;
 	private long lastMoveCall = Time.alert(60);
 
-	public Enemy(int x, int y, int health, int strength, int speed) {
+	public Enemy(int x, int y, int health, int strength, float speed) {
 		super(x, y, 1, 1);
 
 		addCollisionBox(new HitBox(this, -12.5f, -12.5f, 25, 25, true));
